db calls for addTransaction - unfinished. Dustin

+++ app/src/main/java/de/fh/mae/md2/app/repository/TransactionRepository.java
@@ -80,7 +80,7 @@ public class TransactionRepository extends AndroidViewModel {
         return list;
     }
 
-    public Transaction getTransactionById(final int id) {
+    public Transaction getTransactionById(final long id) {
         try{
             new  AsyncTask<Void, Void, Void>(){
                 @Override
